P0Abrufen/requirements.txt
emit
mysql-connector-python
logging
datetime
black
feedparser
argparse